+++ src/redux/version/versionMiddleware.ts
@@ -3,8 +3,7 @@ import { createAsyncThunk } from '@reduxjs/toolkit';
 import { api } from '@/api/api';
 import { onOpenConnection } from '@/redux/connection/connectionSlice';
 import { AppStartListening } from '@/redux/listenerMiddleware';
-
-import { OpenSpaceVersion } from './versionSlice';
+import { OpenSpaceVersion } from '@/types/types';
 
 export const getVersion = createAsyncThunk('version/getVersion', async () => {
   const topic = api.startTopic('version', {});
