+++ src/windowmanagement/HelpMenu.tsx
@@ -1,7 +1,13 @@
-import { Anchor, Button, Grid, Image, Menu, Modal, Text, Title } from '@mantine/core';
+import { Button, Menu } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 
-import { useOpenSpaceApi } from '@/api/hooks';
+import {
+  useGetBoolPropertyValue,
+  useGetIntPropertyValue,
+  useGetStringPropertyValue,
+  useOpenSpaceApi
+} from '@/api/hooks';
+import { About } from '@/components/About/About';
 import {
   BookIcon,
   ConsoleIcon,
@@ -11,34 +17,21 @@ import {
   OpenInBrowserIcon,
   RouteIcon
 } from '@/icons/icons';
-import { useAppSelector } from '@/redux/hooks';
 
 export function HelpMenu() {
-  const [opened, { open, close }] = useDisclosure(false);
   const luaApi = useOpenSpaceApi();
-  const openSpaceVersion = useAppSelector((state) => state.version.openSpaceVersion);
+  const [opened, { open, close }] = useDisclosure(false);
+  const [isConsoleVisible, setIsConsoleVisible] =
+    useGetBoolPropertyValue('LuaConsole.IsVisible');
+  const [portProperty] = useGetIntPropertyValue('Modules.WebGui.Port');
+  const [addressProperty] = useGetStringPropertyValue('Modules.WebGui.Address');
 
-  async function toggleLuaConsole() {
-    if (!luaApi) {
-      return;
-    }
-    const visible = (await luaApi.propertyValue('LuaConsole.IsVisible')) as
-      | boolean
-      | undefined;
-    luaApi.setPropertyValueSingle('LuaConsole.IsVisible', !visible);
+  function toggleLuaConsole() {
+    setIsConsoleVisible(!isConsoleVisible);
   }
 
-  async function openGuiInBrowser() {
-    if (!luaApi) {
-      return;
-    }
-    const portProperty = (await luaApi.propertyValue('Modules.WebGui.Port')) as
-      | number
-      | undefined;
+  function openGuiInBrowser() {
     const port = portProperty ?? 4670; // TODO: Change this to correct port when hosted
-    const addressProperty = (await luaApi.propertyValue('Modules.WebGui.Address')) as
-      | string
-      | undefined;
     const address = addressProperty ?? 'localhost';
 
     const link = `http://${address}:${port}`;
@@ -49,49 +42,9 @@ export function HelpMenu() {
     luaApi?.toggleShutdown();
   }
 
-  function osVersionNumber(): React.JSX.Element {
-    if (!openSpaceVersion) {
-      return <Text>Fetching OpenSpace version...</Text>;
-    }
-
-    function formatVersion(version: {
-      major: number;
-      minor: number;
-      patch: number;
-    }): string {
-      return version.major !== 255 && version.minor !== 255 && version.patch !== 255
-        ? `${version.major}.${version.minor}.${version.patch}`
-        : 'Custom';
-    }
-
-    return <Text>OpenSpave version: {formatVersion(openSpaceVersion)}</Text>;
-  }
-
   return (
     <>
-      <Modal opened={opened} onClose={close} title={'About'} size={'40%'}>
-        <Grid>
-          <Grid.Col span={4}>
-            <Image src={'openspace-logo.png'} alt={'OpenSpace logo'} w={'100%'} />
-          </Grid.Col>
-          <Grid.Col span={8}>
-            <Title order={1}>OpenSpace</Title>
-
-            <Text>
-              OpenSpace is open source interactive data visualization software designed to
-              visualize the entire known universe and portray our ongoing efforts to
-              investigate the cosmos.
-            </Text>
-            {osVersionNumber()}
-            <Text>
-              &copy; 2014 - {new Date().getFullYear()} OpenSpace Development Team
-            </Text>
-            <Anchor href={'https://www.openspaceproject.com/'} target={'_blank'}>
-              openspaceproject.com
-            </Anchor>
-          </Grid.Col>
-        </Grid>
-      </Modal>
+      <About opened={opened} close={close} />
 
       <Menu position={'bottom-start'} offset={4} withArrow arrowPosition={'center'}>
         <Menu.Target>
